Support AWS Signature Version 4 authorization header The need is to use logstash together with Amazon ElasticSearch Service. To provide this support, I added the following parameters in the appender configuration:
<UseAWS4Signer>true/false</UseAWS4Signer>
<AWS4SignerRegion>aws es region</AWS4SignerRegion>
<AWS4SignerAccessKey>aws user access key</AWS4SignerAccessKey>
<AWS4SignerSecretKey>aws user secret key</AWS4SignerSecretKey>

The idea is to calculate the correct signature for requests
according to the rules can be found here: http://docs.aws.amazon.com/general/latest/gr/sigv4_signing.html

I tried to keep changes the simplest possible, but I confess I do not like the way the code that generates the header
was after the changes. It seems that it could be more polymorphic, but this would change a little too much of the code structure.